% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Papal.R
\docType{package}
\name{Papal}
\alias{Papal}
\title{Data on Papal Visits and the country's human rights score for that year.}
\format{
A dataframe with 9329 rows and 5 columns
\describe{
\item{country}{Country name}
\item{year}{The year referred to}
\item{visit}{Binary indicator: 1 if a Pope visited and 0 if not}
\item{name_pope}{Name of the Pope who visited (blank if visit is 0)}
\item{human_rights_score}{The country’s “latent human rights” score for that year}
}
}
\source{
\url{https://tinyletter.com/data-is-plural/letters/data-is-plural-2020-09-02-edition}
}
\usage{
Papal
}
\description{
This package contains information on human rights scores given that
there was a Papal visit or not in a certain year.

A data set containing human rights scores per country per year
given there was a Papal visit or not
}
\keyword{datasets}
